+++ vite-electron.config.ts
@@ -1,31 +1,61 @@
-import { defineConfig } from 'vite'
+import { defineConfig } from 'electron-vite'
 import { resolve } from 'path'
 
 export default defineConfig({
-  build: {
-    lib: {
-      entry: 'electron/main.ts',
-      formats: ['cjs'],
-      fileName: 'main',
-    },
-    rollupOptions: {
-      external: ['electron'],
-      output: {
-        globals: {
-          electron: 'require("electron")',
-        },
+  main: {
+    build: {
+      // @ts-ignore
+      lib: {
+        entry: resolve(__dirname, 'electron/main.ts'),
+        formats: ['cjs'],
+        fileName: 'main'
       },
+      // @ts-ignore
+      rollupOptions: {
+        external: ['electron'],
+        output: {
+          globals: {
+            electron: 'require("electron")'
+          }
+        }
+      }
     },
-  },
-  resolve: {
-    alias: {
-      '@': resolve(__dirname, 'src'),
+    resolve: {
+      alias: {
+        '@': resolve(__dirname, 'src')
+      }
     },
+    define: {
+      'process.env': '{}'
+    },
+    ssr: {
+      noExternal: ['path', 'child_process', 'fs', 'os']
+    }
   },
-  define: {
-    'process.env': '{}',
+  preload: {
+    build: {
+      // @ts-ignore
+      lib: {
+        entry: resolve(__dirname, 'electron/preload.ts'),
+        formats: ['cjs'],
+        fileName: 'preload'
+      },
+      // @ts-ignore
+      rollupOptions: {
+        external: ['electron'],
+        output: {
+          globals: {
+            electron: 'require("electron")'
+          }
+        }
+      }
+    }
   },
-  ssr: {
-    noExternal: ['path', 'child_process', 'fs', 'os'],
-  },
+  renderer: {
+    resolve: {
+      alias: {
+        '@': resolve(__dirname, 'src')
+      }
+    }
+  }
 })